chore: more prettier code
